requirements.txt
# nose
# sphinx

# ring_doorbell==0.6.2
ring_doorbell==0.7.2
python-decouple
Flask
# decouple
functions-framework==3.*
google-cloud-secret-manager
